ref: Move releases sub-components into correct folder
Since GroupList, CommitAuthorStats and ReleaseProjectStatSparkLine
components are only used within releases we should move them into
the releases views folder. Since they are not broadly reused they
shouldn't be classified as components.
