+++ src/sentry/static/sentry/app/views/releases/detail/organization/releaseOverview.jsx
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types';
 
 import SentryTypes from 'app/sentryTypes';
 import LastCommit from 'app/components/lastCommit';
-import CommitAuthorStats from 'app/components/commitAuthorStats';
-import ReleaseProjectStatSparkline from 'app/components/releaseProjectStatSparkline';
+
 import RepositoryFileSummary from 'app/components/repositoryFileSummary';
 import AsyncComponent from 'app/components/asyncComponent';
 import {t} from 'app/locale';
@@ -13,6 +12,8 @@ import {getFilesByRepository} from '../shared/utils';
 import ReleaseDeploys from '../shared/releaseDeploys';
 import ReleaseEmptyState from '../shared/releaseEmptyState';
 import ReleaseIssues from '../shared/releaseIssues';
+import CommitAuthorStats from '../shared/commitAuthorStats';
+import ReleaseProjectStatSparkline from '../shared/releaseProjectStatSparkline';
 
 export default class OrganizationReleaseOverview extends AsyncComponent {
   static propTypes = {
