+++ src/sentry/static/sentry/app/views/releases/detail/project/releaseNewEvents.jsx
@@ -3,9 +3,10 @@ import React from 'react';
 import {Link} from 'react-router';
 
 import SentryTypes from 'app/sentryTypes';
-import GroupList from 'app/components/groupList';
 import {t} from 'app/locale';
 import withEnvironmentInQueryString from 'app/utils/withEnvironmentInQueryString';
+
+import GroupList from '../shared/groupList';
 
 class ReleaseNewEvents extends React.Component {
   static propTypes = {
