+++ src/sentry/static/sentry/app/views/releases/detail/shared/commitAuthorStats.jsx
