+++ src/sentry/static/sentry/app/views/releases/detail/shared/releaseProjectStatSparkline.jsx
